+++ CHANGELOG.md
@@ -18,6 +18,7 @@ See [github commits](https://github.com/TA-Lib/ta-lib/commits) for complete list
 ### Fixed
 - (#57) Missing TA_GetVersionString function in Windows DLL. Thanks @Youngv !
 - (#98) TRIX and NATR returned wrong values when startIdx > lookback, and a non-zero unstable period changed IMI's summation window.
+- (#99) BBANDS with `TA_MAType_MAMA` and a period >= 34 returned a misaligned middle band.
 
 ## [0.7.1] 2026-07-03
 ### Added